Move window state things to `settings`
I think it makes more sense to put them in `settings` as they make
intense use of the `Settings` enum.

It also cleans `app` a bit, because it is not using `settings` directly
now.
